Make SpellCheck#CheckErrors() take an additional a:types argument to support filtering for certain spell error types
Add loop and iterate until an corresponding spell error type is encountered or an exit condition is met.
Always position the cursor at the beginning of the checked range. This avoids the workaround in s:GotoNextSpellError() and lets us :set nowrapscan, to avoid endless iteration.
Go to the first misspelling in the passed range (if any) instead of the first overall. I think this is more DWIM.
Extract SpellCheck#NoErrorsFoundMessage() and create SpellCheck#GetTypes() for reuse.