Add configuration for highlighting of the error word and context in the quickfix window
Introduce additional g:SpellCheck_SpellWordPattern because spell checking doesn't exactly use the 'iskeyword' option, and we have to emulate it.
FIX: Quickfix mappings are gone when closing and reopening the quickfix window (:cclose | copen), because a new scratch buffer is used, but the autocmd BufRead quickfix has been cleared.
Only clear that autocmd when a different quickfix source is used. (And then also turn off the plugin's additional syntax highlighting.)