Skip spell checking in the quickfix window / location list window itself
This mostly applies to using :window SpellCheck when the quickfix window is already open. It might still make sense (somehow) to check quickfix output from another tool, so just make :SpellCheck apply to the quickfix window and :SpellLCheck to location lists.

Do the check before SpellCheck#CheckEnabledSpelling() so that we don't turn on spell checking in the quickfix window.

Closes #2